+++ app/i18n/es.js
@@ -1,0 +1,8 @@
+"use strict";
+
+export const myEs = {
+  welcome: "Bienvenido a la API REST (Gis-Service)",
+  profileRequired: "Se requiere un perfil.",
+  profileInvalid: "El perfil no es válido.",
+  coordinatesInvalid: "Las coordenadas no son válidas.",
+};