+++ app/i18n/index.js
@@ -2,5 +2,6 @@
 
 import { myDe } from "./de.js";
 import { myEn } from "./en.js";
+import { myEs } from "./es.js";
 
-export { myDe, myEn };
+export { myDe, myEn, myEs };